pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nexus-ai"
version = "2.0.0"
description = "Multi-agent AI with 6-source live verification, debate convergence, and production security. Beats OpenClaw on accuracy, safety, and real-time data."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
keywords = ["ai","agents","multi-agent","security","real-time","llm"]
authors = [{name = "Nexus AI"}]
classifiers = [
    "Programming Language :: Python :: 3.11",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "python-multipart>=0.0.9",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "slowapi>=0.1.9",
    "keyring>=25.2.0",
    "aiosqlite>=0.20.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "python-dotenv>=1.0.1",
    "httpx>=0.27.0",
    "structlog>=24.2.0",
    "aiofiles>=23.2.1",
    "apscheduler>=3.10.4",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.3",
    "lxml>=5.2.0",
]

[project.optional-dependencies]
full = [
    "playwright>=1.44.0",
    "faiss-cpu>=1.8.0",
    "sentence-transformers>=3.0.0",
    "huggingface-hub>=0.23.0",
    "groq>=0.9.0",
    "spotipy>=2.23.0",
    "google-auth>=2.29.0",
    "google-api-python-client>=2.131.0",
    "python-telegram-bot>=21.3",
    "presidio-analyzer>=2.2.354",
    "presidio-anonymizer>=2.2.354",
    "streamlit>=1.35.0",
    "openai-whisper>=20231117",
    "gtts>=2.5.0",
]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.7",
    "pytest-cov>=5.0.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

[project.scripts]
nexus = "src.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "config*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.ruff]
line-length = 100
target-version = "py311"


